refactor: extract validation and compose resolution from infra to zero (#8364)
Move validateAgentSession, validateCheckpoint, resolveStartRunCompose and
their helper functions from infra/run/run-service.ts to a new
zero/zero-run-validation.ts. Rebased on latest main which already extracted
concurrency control to zero-run-policy.ts.
