+++ turbo/apps/web/src/lib/infra/run/index.ts
@@ -4,11 +4,9 @@
  */
 
 export {
-  validateAgentSession,
   startRun,
   createRunRecord,
   buildAndDispatchRun,
-  resolveStartRunCompose,
   loadCompose,
   markRunFailed,
   registerCallbacks,
